fix day in model

+++ lib/models/weather_forecast_daily.dart
@@ -225,7 +225,7 @@ class FeelsLike {
   FeelsLike({this.day, this.night, this.eve, this.morn});
 
   FeelsLike.fromJson(Map<String, dynamic> json) {
-    day = json['day'];
+    day = json['day'].toDouble();
     night = json['night'];
     eve = json['eve'];
     morn = json['morn'].toDouble();
